Only create new token if old token is not valid anymore

+++ src/Security/EmailConfirmation/ConfirmationManager.php
@@ -41,18 +41,18 @@ class ConfirmationManager {
     public function newConfirmation(User $user, string $email): void {
         $confirmation = $this->repository->findOneByUser($user);
 
-        if($confirmation !== null) {
+        if($confirmation !== null && $confirmation->getValidUntil() <= $this->dateHelper->getNow()) {
             $this->repository->remove($confirmation);
+        } else {
+            $confirmation = (new EmailConfirmation())
+                ->setUser($user)
+                ->setEmailAddress($email)
+                ->setValidUntil($this->dateHelper->getNow()->modify(static::Lifetime));
+
+            do {
+                $confirmation->setToken(bin2hex(openssl_random_pseudo_bytes(64)));
+            } while($this->repository->findOneByToken($confirmation->getToken()) !== null);
         }
-
-        $confirmation = (new EmailConfirmation())
-            ->setUser($user)
-            ->setEmailAddress($email)
-            ->setValidUntil($this->dateHelper->getNow()->modify(static::Lifetime));
-
-        do {
-            $confirmation->setToken(bin2hex(openssl_random_pseudo_bytes(64)));
-        } while($this->repository->findOneByToken($confirmation->getToken()) !== null);
 
         $this->repository->persist($confirmation);
